+++ opentelemetry-user-events-logs/src/logs/mod.rs
@@ -1,4 +1,6 @@
 mod exporter;
 mod processor;
 
+#[cfg(feature = "experimental_eventname_callback")]
+pub use exporter::EventNameCallback;
 pub use processor::{Processor, ProcessorBuilder};